백업

+++ app-public-api/src/main/java/co/kurrant/app/public_api/dto/order/UpdateCart.java
@@ -2,8 +2,10 @@ package co.kurrant.app.public_api.dto.order;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class UpdateCart {
-    private Integer dailyFoodId;
+    private BigInteger dailyFoodId;
     private Integer count;
 }
