+++ domain-food/src/main/java/co/dalicious/domain/food/entity/Food.java
@@ -10,6 +10,8 @@ import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @DynamicInsert
 @DynamicUpdate
@@ -23,7 +25,7 @@ public class Food {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     @Comment("ID")
-    private Integer id;
+    private BigInteger id;
 
     @Column(name = "name")
     @Comment("식품 이름")
@@ -31,7 +33,7 @@ public class Food {
 
     @Column(name = "price")
     @Comment("가격")
-    private Integer price;
+    private BigDecimal price;
 
     @Column(name = "img")
     @Comment("이미지 경로")
@@ -39,7 +41,7 @@ public class Food {
 
     @Column(name = "discounted_rate")
     @Comment("할인율")
-    private Double discountedRate;
+    private BigDecimal discountedRate;
 
     @Column(name = "e_spicy")
     @Comment("맵기정도")
@@ -56,7 +58,7 @@ public class Food {
     private String description;
 
     @Builder
-    Food(Integer id, String name, Integer price, String img, Makers makers, String description, Double discountedRate, Spicy spicy){
+    Food(BigInteger id, String name, BigDecimal price, String img, Makers makers, String description, BigDecimal discountedRate, Spicy spicy){
         this.id = id;
         this.name = name;
         this.price = price;
