+++ domain-food/src/main/java/co/dalicious/domain/food/repository/QOriginRepository.java
@@ -5,6 +5,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static co.dalicious.domain.food.entity.QOrigin.origin;
@@ -14,7 +15,7 @@ import static co.dalicious.domain.food.entity.QOrigin.origin;
 public class QOriginRepository {
     public final JPAQueryFactory queryFactory;
 
-    public List<Origin> findByFoodId(Integer foodId) {
+    public List<Origin> findByFoodId(BigInteger foodId) {
               return queryFactory.selectFrom(origin)
                 .where(origin.food.id.eq(foodId))
                 .fetch();
