+++ domain-makers/src/main/java/co/dalicious/domain/makers/entity/Makers.java
@@ -7,6 +7,7 @@ import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @DynamicInsert
 @DynamicUpdate
@@ -18,12 +19,12 @@ public class Makers {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private BigInteger id;
     @Column(name = "name")
     private String name;
 
     @Builder
-    Makers(Integer id, String name){
+    Makers(BigInteger id, String name){
         this.id = id;
         this.name = name;
     }
